nextjs resolver / dynamic route
